Assignment-2/mnt/user-data/outputs/codebase-genius/requirements.txt
# Core Framework
jaclang>=0.7.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# HTTP and API
requests>=2.31.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Git operations
GitPython>=3.1.40

# Code analysis (AST is built-in, but adding optional enhancers)
typing-extensions>=4.8.0

# Utilities
python-dateutil>=2.8.2
pathlib>=1.0.1

# Optional: For enhanced features
# tree-sitter>=0.20.0  # Optional: Advanced parsing
# openai>=1.0.0  # Optional: LLM integration
# anthropic>=0.7.0  # Optional: Claude API
